03Febv2

+++ controller/feedController.js
@@ -57,11 +57,9 @@ export const getAnnouncements = async (req, res) => {
       a.date_to,
       a.active,
       (SELECT COUNT(*) FROM v4.shared_comments WHERE relation_id = a.row_id AND relation_type = 'announcements') as comment_count,
-       -- ✅ Add view count
-      (SELECT COUNT(*) FROM v4.announcement_views WHERE announcement_id = a.row_id) as view_count,
-      -- ✅ Check if current user has viewed
-      EXISTS(SELECT 1 FROM v4.announcement_views WHERE announcement_id = a.row_id AND user_id = $2::uuid) as has_viewed,
-      a.comments_on,
+      (SELECT COUNT(*) FROM v4.announcement_views WHERE announcement_id = a.row_id::integer) as view_count,
+      EXISTS(SELECT 1 FROM v4.announcement_views WHERE announcement_id = a.row_id::integer AND user_id = $2::uuid) as has_viewed,
+       a.comments_on,
       a.created_by,
       to_char(a.created_at AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS"Z"') as created_at,
       u.first_name || ' ' || u.last_name as created_by_name,
